moved middleware and errors to separate dirs
